refactored common functions
